Fixed a bug, then fixed a bug about a bug.
Noticed that the bash prompt over SSH didn't have the neat separator
thing that I had put in. Very disappointed, so fixed the offending
.bash_profile type.

And THEN noticed that after THAT bug was fixed, TRAMP pulak buat
hal. Haih. Tak best betul. So fixed the .bashrc based on a StackOverflow
question that dealt with this issue (TLDR; TRAMP is giant baby, wants
prompt to be ~~just so~~).